# server_code/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: Follow Me Cart API
  description: This API allows controlling the Follow Me Cart.

definitions:
  User:
    type: object
    properties:
      user_id:
        type: integer
        description: Unique identifier for the user
      user_password:
        type: string
        description: Password for the user login
      user_full_name:
        type: string
        description: User's full name

  Empty:
    type: object
    properties:

# Paths for user operations
paths:
  /users:
    get:
      summary: List all users
      description: Retrieves a list of all users in the system.
      responses:
        200:
          description: Successful operation
          schema:
            type: array
            items:
              $ref: "#/definitions/User"

  /users/create:
    post:
      summary: Create a new user
      description: Creates a new user account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/definitions/User"
      parameters:
        - name: User ID
          in: user_id
          required: true
          type: integer
          description: ID of the user to create
          example: 0
        - name: User password
          in: user_password
          required: true
          type: string
          description: Password of the user to create
          example: 'notThisPassword'
        - name: User full name
          in: user_full_name
          required: true
          type: string
          description: User full name of the user to create
          example: 'João do Monte'
      responses:
        200:
          description: User created successfully
          content:
            application/json:
              schema:
              $ref: "#/definitions/Empty"
        409:
          description: Conflict (user_id already exists)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: "User already exists"

  /users/delete:
    post:
      summary: Delete a user
      description: Deletes a specific user by their ID.
      requestBody:
        required: true
        content:
          application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: integer
                    description: Unique identifier for the user
      responses:
        200:
          description: User deleted successfully
          content:
            application/json:
              schema:
              $ref: "#/definitions/Empty"
        404:
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: "User not found"

# Global parameters (optional)
# servers:
# - url: https://your-api-server.com/

# Security (optional)
# security:
# - api_key: []